net/ipv6/addrconf: IPv6 tethering enhancement
Added new procfs flag to toggle the automatic addition of prefix
routes on a per device basis. The new flag is accept_ra_prefix_route.
Defaults to 1 as to not break existing behavior.
*
Added this to fix tethering on the v510.  Doesnt fix it, but does seem to be beneficial. probably effect.